+++ libs/wiring_backend/src/host_client_config.rs
@@ -5,8 +5,8 @@
 pub const LOBBY_LIST_SIZE: usize = 10;
 pub const ACK_TIMEOUT_MILLIS: u64 = 14_000;
 
-pub const MAX_LOBBY_PLAYERS: usize = 4;
-pub const MAX_LOBBY_WATCHERS: usize = 4;
-pub const MIN_PLAYERS_TO_LAUNCH: usize = 1;
+pub const MAX_LOBBY_PLAYERS: u16 = 4;
+pub const MAX_LOBBY_WATCHERS: u16 = 4;
+pub const MIN_PLAYERS_TO_LAUNCH: u16 = 1;
 
 //-------------------------------------------------------------------------------------------------------------------
